DDF-2195 Create Registry Rest Endpoint (#866)
* DDF-2195 Registry Rest Endpoint

Changing the registry-report-view module to registry-rest-endpoint
Moving the buildRegistryMap logic to a RegistryReportHelper
Adding publish to the rest endpoint  POST /{registryId}/publication/{sourceId}
Adding unpublish to the rest endpoint DELETE /{registryId}/publication/{sourceId}
Updating tests

* DDF-2195 Registry Rest Endpoint Addressing PR Comments

Renaming RegistryReportHelper to RegistryReportMapBuilder
Refactoring paramenter validation for publish and unpublish
Fixing error response for internal errors
Updating tests

* DDF-2195 Registry Rest Endpoint

Moving registry-app to the bottom of the modules list
Removing the version for the commons-lang3 dependency in the registry-rest-endpoint pom.
To use the commons-lang3 dependency that is now under catalog's dependency management

* DDF-2195 Registry Rest Endpoint Addressing PR Comments

Reverting the contents of report.hbt to what was in master. The file remained moved.

(cherry picked from commit 6053219f8194097fc84c04b41e5bc5ed66dc55c9)
