--- catalog/spatial/registry/registry-report-viewer/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,46 +1,0 @@
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
- -->
-<blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns:jaxrs="http://cxf.apache.org/blueprint/jaxrs"
-           xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xsi:schemaLocation="
-  http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
-
-    <bean id="resourceMapping"
-          class="org.ops4j.pax.web.extender.whiteboard.runtime.DefaultResourceMapping">
-        <property name="alias" value="/services/registries/assets"/>
-        <property name="path" value="/assets"/>
-    </bean>
-
-    <bean id="RegistryReportViewer"
-          class="org.codice.ddf.registry.report.viewer.RegistryReportViewer">
-        <property name="federationAdminService" ref="federationAdminService"/>
-    </bean>
-
-    <jaxrs:server id="RegistryReportViewerService" address="/registries">
-        <jaxrs:serviceBeans>
-            <ref component-id="RegistryReportViewer"/>
-        </jaxrs:serviceBeans>
-        <jaxrs:extensionMappings>
-            <entry key="html" value="text/html"/>
-        </jaxrs:extensionMappings>
-    </jaxrs:server>
-
-    <reference id="federationAdminService"
-               interface="org.codice.ddf.registry.federationadmin.service.FederationAdminService"
-               availability="mandatory"/>
-
-    <service id="resources" ref="resourceMapping"
-             interface="org.ops4j.pax.web.extender.whiteboard.ResourceMapping"/>
-</blueprint>
