+++ catalog/spatial/registry/registry-rest-endpoint/src/main/resources/assets/img/glyphicons-halflings-white.png
