+++ catalog/spatial/registry/registry-rest-endpoint/src/main/resources/templates/report.hbt
@@ -1,0 +1,206 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta content="text/html; charset=windows-1252" http-equiv="content-type" />
+    <!-- styles -->
+    <link href="/services/registries/assets/css/bootstrap.min.css" rel="stylesheet" />
+    <style type="text/css">
+        body {
+        background-color: {
+        {
+            background
+        }
+        }
+        ;
+        color: {
+        {
+            foreground
+        }
+        }
+        ;
+        margin-left: 15px;
+        }
+
+        .scrollable-list {
+            height: auto;
+            max-height: 600px;
+            overflow: auto;
+        }
+
+        .spinner {
+            height: 100px;
+            position: relative;
+        }
+
+        .banner-container {
+            text-align: center;
+            height: 20px;
+        }
+
+        .header-layout,
+        .footer-layout,
+        #content {
+            position: fixed;
+            right: 0;
+            left: 0;
+        }
+
+        .header-layout {
+            top: 0;
+        }
+
+        .footer-layout {
+            bottom: 0;
+        }
+
+        #content {
+            top: 0px;
+            bottom: 0px;
+            overflow: auto;
+        }
+
+        .jumbotron {
+            background-color: rgba(0, 0, 0, 0.25);
+        }
+
+        .panel {
+            color: black;
+        }
+
+        .table {
+            margin-left: 50px;
+            border: 1px solid #ddd;
+            max-width: 2000px;
+        }
+
+        h3 {
+            margin-left: 25px
+        }
+
+        h4 {
+            margin-left: 50px
+        }
+    </style>
+    <link href="/services/registries/assets/css/bootstrap-theme.min.css" rel="stylesheet" />
+</head>
+
+<body>
+<h1>{{Name}}</h1>
+<table class="table table-striped">
+    <tbody>
+    {{#each General}}
+        <tr>
+            <td width="250 px">
+                <strong>{{@key}}</strong>
+            </td>
+            <td>{{this}}</td>
+        </tr>
+    {{/each}}
+    </tbody>
+</table>
+<h2>Services</h2>
+{{#each Services}}
+    <h3>{{@key}}</h3>
+    <table class="table table-striped">
+        <tbody>
+        {{#each this.Properties}}
+            <tr>
+                <td width="250 px">
+                    <strong>{{@key}}</strong>
+                </td>
+                <td>{{this}}</td>
+            </tr>
+        {{/each}}
+        </tbody>
+    </table>
+    {{#each this.Bindings}}
+        <h4>{{@key}}</h4>
+        <table class="table table-striped ">
+            <tbody>
+            {{#each this}}
+                <tr>
+                    <td width="250 px">
+                        <strong>{{@key}}</strong>
+                    </td>
+                    <td>{{this}}</td>
+                </tr>
+            {{/each}}
+            </tbody>
+        </table>
+    {{/each}} {{/each}}
+<h2>Organizations</h2>
+{{#each Organizations}}
+    <h3>{{@key}}</h3>
+    {{#each this.ContactInfo}}
+        <h4>{{@key}}</h4>
+        <table class="table table-striped ">
+            <tbody>
+            {{#each this}}
+                <tr>
+                    <td>{{this}}</td>
+                </tr>
+            {{/each}}
+            </tbody>
+        </table>
+    {{/each}}
+    <table class="table table-striped ">
+        <tbody>
+        {{#each this.CustomSlots}}
+            <tr>
+                <td width="250 px">
+                    <strong>{{@key}}</strong>
+                </td>
+                <td>{{this}}</td>
+            </tr>
+        {{/each}}
+        </tbody>
+    </table>
+{{/each}}
+<h2>Contacts</h2>
+{{#each Contacts}}
+    <h3>{{@key}}</h3>
+    {{#each this.ContactInfo}}
+        <h4>{{@key}}</h4>
+        <table class="table table-striped ">
+            <tbody>
+            {{#each this}}
+                <tr>
+                    <td>{{this}}</td>
+                </tr>
+            {{/each}}
+            </tbody>
+        </table>
+    {{/each}}
+    <table class="table table-striped ">
+        <tbody>
+        {{#each this.CustomSlots}}
+            <tr>
+                <td width="250 px">
+                    <strong>{{@key}}</strong>
+                </td>
+                <td>{{this}}</td>
+            </tr>
+        {{/each}}
+        </tbody>
+    </table>
+{{/each}}
+<h2>Collections</h2>
+{{#each Collections}}
+    <h3>{{@key}}</h3>
+    <table class="table table-striped ">
+        <tbody>
+        {{#each this}}
+            <tr>
+                <td width="250 px">
+                    <strong>{{@key}}</strong>
+                </td>
+                <td>{{this}}</td>
+            </tr>
+        {{/each}}
+        </tbody>
+    </table>
+{{/each}}
+</body>
+
+</html>
