+++ catalog/spatial/registry/registry-rest-endpoint/src/test/resources/csw-bare-registry-package.xml
