+++ catalog/spatial/registry/registry-rest-endpoint/src/test/resources/csw-sparse-registry-package.xml
